version 4.2
some bugs fixed. Especially if a voice does not have the required Name attribute it will be ignored in the dual voice secondary voice dialog. Moreover, if the previously selected voices are uninstalled the dual voice can set use default voice instead of them. Furthermore, if the secondary voice leads to an error the dual voice could handle the error by using the primary voice instead of it.